move NotificationsProviders to features/NotificationsPanel
