+++ common/v2/features/Dashboard/Dashboard.tsx
@@ -7,8 +7,8 @@ import { useDevMode } from 'v2/services';
 import { AccountContext, AddressBookContext } from 'v2/services/Store';
 import { translate, translateRaw } from 'translations';
 import { AccountList, BannerAd, Desktop, Mobile } from 'v2/components';
+import { NotificationsPanel } from '../NotificationsPanel';
 import { ActionTile, TokenList, WalletBreakdown, RecentTransactionList } from './components';
-import { NotificationsPanel } from './NotificationsPanel';
 import { actions } from './constants';
 import './Dashboard.scss';
