--- common/v2/features/Dashboard/NotificationsPanel/index.ts
@@ -1,1 +1,0 @@
-export { default as NotificationsPanel } from './NotificationsPanel';
