+++ common/v2/features/NotificationsPanel/components/GetHardwareWalletNotification.tsx
