+++ common/v2/features/NotificationsPanel/components/OnboardingResponsibleNotification.tsx
