+++ common/v2/features/NotificationsPanel/components/WalletAddedNotification.tsx
