+++ common/v2/features/NotificationsPanel/components/WalletCreatedNotification.tsx
