+++ common/v2/features/NotificationsPanel/components/WalletNotAddedNotification.tsx
