+++ common/v2/features/NotificationsPanel/index.ts
@@ -1,0 +1,3 @@
+export { default as NotificationsPanel } from './NotificationsPanel';
+export { NotificationTemplates, notificationsConfigs } from './constants';
+export { NotificationsContext, NotificationsProvider } from './NotificationsProvider';
