--- common/v2/providers/NotificationsProvider/index.ts
@@ -1,3 +1,0 @@
-export * from './types';
-export * from './constants';
-export * from './NotificationsProvider';
